Fix and simplify the common relax workflow launch scripts
Since the recent change, an inputs generator needs to be passed the
corresponding process class upon construction. However, it is a lot
easier, and the suggested method, to retrieve an instance from the
`get_inputs_generator` method exposed on the workchain. Since the
workchains now have proper entry points, they can be loaded through
their corresponding entry point.
